+++ src/mos.h
@@ -156,7 +156,7 @@ UINT8	fat_EOF(FIL * fp);
 UINT8	fat_size(FIL * fp, DWORD * size);
 UINT8	fat_error(FIL * fp);
 UINT8	fat_lseek(FIL * fp, DWORD * offset);
-int		fat_getfree(const TCHAR * path, DWORD * clusters, DWORD * clusterSize);
+UINT8	fat_getfree(const TCHAR * path, DWORD * clusters, DWORD * clusterSize);
 
 UINT8	wait_VDP(UINT8 mask);
